feat(devxrl): replace /devxrl page with new one-pager design
- Rebuild DevXRL.vue from updated HTML source
- Add pipeline diagram with activation and conversion funnel trapezoids
- Add hero scroll-hint and TRL framework badges (TRL, DevXRL, G2MRL)
- Derive nav dot count from sectionList array instead of hardcoded 10
- Consolidate DevXRL CSS variables under --dxrl-* prefix scoped on .devxrl
- Prefix all DevXRL class names with devxrl-
- Scope noise overlay to html.devxrl-page
- Update CLAUDE.md page-by-page reference and CHANGELOG.md
