+++ ROADMAP.md
@@ -95,4 +95,4 @@ Items move here when finished, with the date and any relevant notes.
 
 ---
 
-*Last updated: 2026-03-28*
+*Last updated: 2026-04-19*
